works with motor now, ends up running at a constant rpm controlled by power supply voltage, not sure of cause

+++ projects/mci/src/motor_can.c
@@ -44,7 +44,6 @@ static inline uint8_t pack_right_shift_u32(uint32_t value, uint8_t shift, uint8_
 int motor_can_drive_command_pack(uint8_t *dst_p, const MotorCanDriveCommand *src_p, size_t size) {
   uint32_t motor_current;
   uint32_t motor_velocity;
-  
 
   if (size < 8u) {
     return (-EINVAL);
@@ -63,6 +62,6 @@ int motor_can_drive_command_pack(uint8_t *dst_p, const MotorCanDriveCommand *src
   dst_p[6] |= pack_right_shift_u32(motor_current, 16u, 0xffu);
   dst_p[7] |= pack_right_shift_u32(motor_current, 24u, 0xffu);
 
-  LOG_DEBUG("motor current %d motor velocity %d???\n", (int)motor_current, (int)motor_velocity);
+  LOG_DEBUG("motor current %d motor velocity %d???\n", (int)((uint32_t)motor_current), (int)((uint32_t)motor_velocity));
   return (8);
 }
